# export/api.yaml
openapi: 3.0.0
paths:
  /auth/login:
    get:
      operationId: loginUser
      parameters:
        - name: username
          required: true
          in: query
          schema:
            type: string
        - name: password
          required: true
          in: query
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenDto"
      tags:
        &a1
        - Auth
  /auth/admin/login:
    get:
      operationId: loginAdmin
      parameters:
        - name: email
          required: true
          in: query
          schema:
            type: string
        - name: password
          required: true
          in: query
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenDto"
      tags: *a1
...
  /auth/google:
    get:
      operationId: redirectToGoogle
      parameters:
        - name: nonce
          required: false
          in: query
          schema:
            type: string
        - name: code_challenge_method
          required: false
          in: query
          schema:
            type: string
        - name: code_challenge
          required: false
          in: query
          schema:
            type: string
        - name: state
          required: false
          in: query
          schema:
            type: string
      responses:
        "302":
          description: ""
      tags: *a1
  /auth/facebook:
    get:
      operationId: redirectToFacebook
      parameters:
        - name: nonce
          required: false
          in: query
          schema:
            type: string
        - name: code_challenge_method
          required: false
          in: query
          schema:
            type: string
        - name: code_challenge
          required: false
          in: query
          schema:
            type: string
        - name: state
          required: false
          in: query
          schema:
            type: string
      responses:
        "302":
          description: ""
      tags: *a1
  /users:
    post:
      operationId: addUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecureUserDto"
      tags:
        &a2
        - Users
    get:
      operationId: getUsersBySearching
      parameters:
        - name: q
          required: true
          in: query
          schema:
            type: string
        - name: teamId
          required: false
          in: query
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecureUserDto"
      tags: *a2
  /users/validation:
    get:
      operationId: getUserDedup
      parameters:
        - name: username
          required: false
          in: query
          schema:
            type: string
        - name: mobilePhone
          required: false
          in: query
          schema:
            type: string
        - name: email
          required: false
          in: query
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserValidationResponseDto"
      tags: *a2
  "/users/{id}":
    get:
      operationId: getUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecureUserDto"
      tags: *a2
  "/users/{id}/detail":
    get:
      operationId: getUserDetailById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    $ref: "#/components/schemas/SecureUserDto"
                  settings:
                    nullable: true
                    $ref: "#/components/schemas/UserSettingsDto"
                  avatars:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserAvatarDto"
      tags: *a2
  "/users/multiple/{ids}":
    get:
      operationId: getMultipleUsersByIds
      parameters:
        - name: ids
          required: true
          in: path
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecureUserDto"
      tags: *a2
  /users/avatars:
    post:
      operationId: postUserAvatar
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserAvatarDto"
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAvatarDto"
      tags: *a2
  "/users/{id}/avatars":
    get:
      operationId: getUserAvatarsByUserId
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: gameId
          required: true
          in: query
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserAvatarDto"
      tags: *a2
  "/users/avatars/{id}":
    put:
      operationId: putUserAvatarById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserAvatarDto"
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAvatarDto"
      tags: *a2
    delete:
      operationId: daleteUserAvatarById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/"
      tags: *a2
...
  "/admin/users/{id}":
    get:
      operationId: get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    put:
      operationId: update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentDto"
      responses:
        "200":
          description: ""
    delete:
      operationId: delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
...
info:
  title: FyTy API
  description: ""
  version: "1.0"
  contact: {}
...
servers:
  - url: https://api.fyty-esport.com
    description: null
    variables: null
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    AccessTokenDto:
      type: object
      properties:
        accessToken:
          type: string
      required:
        - accessToken
    CreateUserDto:
      type: object
      properties:
        id:
          type: string
          format: cuid
        username:
          type: string
        updatedUsernameAt:
          nullable: true
        mobilePhone:
          type: string
        updatedMobilePhoneAt:
          nullable: true
        password:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
          nullable: true
        updatedEmailAt:
          nullable: true
        bio:
          type: string
        portraitUrl:
          type: string
        coverUrl:
          type: string
        isVerified:
          type: boolean
        isDeactivated:
          type: boolean
        facebookId:
          type: string
          nullable: true
        googleId:
          type: string
          nullable: true
        lineId:
          type: string
          nullable: true
        lineToken:
          type: string
          nullable: true
        firstLoginAt:
          nullable: true
        lastLoginAt:
          nullable: true
        createdAt: {}
        updatedAt: {}
      required:
        - username
        - mobilePhone
        - password
        - displayName
    SecureUserDto:
      type: object
      properties:
        id:
          type: string
          format: cuid
        username:
          type: string
        updatedUsernameAt:
          nullable: true
        mobilePhone:
          type: string
        updatedMobilePhoneAt:
          nullable: true
        displayName:
          type: string
        email:
          type: string
          format: email
          nullable: true
        updatedEmailAt:
          nullable: true
        bio:
          type: string
        portraitUrl:
          type: string
        coverUrl:
          type: string
        isVerified:
          type: boolean
        isDeactivated:
          type: boolean
        facebookId:
          type: string
          nullable: true
        googleId:
          type: string
          nullable: true
        lineId:
          type: string
          nullable: true
        lineToken:
          type: string
          nullable: true
        firstLoginAt:
          nullable: true
        lastLoginAt:
          nullable: true
        createdAt: {}
        updatedAt: {}
      required:
        - id
        - username
        - mobilePhone
        - displayName
        - bio
        - portraitUrl
        - coverUrl
        - isVerified
        - isDeactivated
        - createdAt
        - updatedAt
    UserValidationResponseDto:
      type: object
      properties:
        username:
          type: boolean
        mobilePhone:
          type: boolean
        email:
          type: boolean
    CreateUserAvatarDto:
      type: object
      properties:
        id:
          type: string
          format: cuid
        inGameId:
          type: string
        characterName:
          type: string
        rank:
          type: string
        ratingScore:
          oneOf:
            - type: number
            - type: string
            - type: object
              properties:
                d:
                  type: array
                  items:
                    type: number
                e:
                  type: number
                s:
                  type: number
                toFixed: {}
              required:
                - d
                - e
                - s
                - toFixed
        gameId:
          type: string
        userId:
          type: string
      required:
        - inGameId
        - rank
        - gameId
        - userId
    UserAvatarDto:
      type: object
      properties:
        id:
          type: string
          format: cuid
        inGameId:
          type: string
        characterName:
          type: string
        rank:
          type: string
        ratingScore:
          oneOf:
            - type: number
            - type: string
            - type: object
              properties:
                d:
                  type: array
                  items:
                    type: number
                e:
                  type: number
                s:
                  type: number
                toFixed: {}
              required:
                - d
                - e
                - s
                - toFixed
        gameId:
          type: string
        userId:
          type: string
      required:
        - id
        - inGameId
        - characterName
        - rank
        - ratingScore
        - gameId
        - userId
    UpdateUserAvatarDto:
      type: object
      properties:
        id:
          type: string
          format: cuid
        inGameId:
          type: string
        characterName:
          type: string
        rank:
          type: string
        ratingScore:
          oneOf:
            - type: number
            - type: string
            - type: object
              properties:
                d:
                  type: array
                  items:
                    type: number
                e:
                  type: number
                s:
                  type: number
                toFixed: {}
              required:
                - d
                - e
                - s
                - toFixed
        gameId:
          type: string
        userId:
          type: string
...
    UpdateAppointmentDto:
      type: object
      properties:
        startAt: {}
        endAt: {}
...